godot: bumped to 4.3.0-beta

+++ drivers/unix/dir_access_unix.cpp
@@ -1,57 +1,53 @@
-/*************************************************************************/
-/*  dir_access_unix.cpp                                                  */
-/*************************************************************************/
-/*                       This file is part of:                           */
-/*                           GODOT ENGINE                                */
-/*                      https://godotengine.org                          */
-/*************************************************************************/
-/* Copyright (c) 2007-2021 Juan Linietsky, Ariel Manzur.                 */
-/* Copyright (c) 2014-2021 Godot Engine contributors (cf. AUTHORS.md).   */
-/*                                                                       */
-/* Permission is hereby granted, free of charge, to any person obtaining */
-/* a copy of this software and associated documentation files (the       */
-/* "Software"), to deal in the Software without restriction, including   */
-/* without limitation the rights to use, copy, modify, merge, publish,   */
-/* distribute, sublicense, and/or sell copies of the Software, and to    */
-/* permit persons to whom the Software is furnished to do so, subject to */
-/* the following conditions:                                             */
-/*                                                                       */
-/* The above copyright notice and this permission notice shall be        */
-/* included in all copies or substantial portions of the Software.       */
-/*                                                                       */
-/* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,       */
-/* EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF    */
-/* MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.*/
-/* IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY  */
-/* CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,  */
-/* TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE     */
-/* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
-/*************************************************************************/
+/**************************************************************************/
+/*  dir_access_unix.cpp                                                   */
+/**************************************************************************/
+/*                         This file is part of:                          */
+/*                             GODOT ENGINE                               */
+/*                        https://godotengine.org                         */
+/**************************************************************************/
+/* Copyright (c) 2014-present Godot Engine contributors (see AUTHORS.md). */
+/* Copyright (c) 2007-2014 Juan Linietsky, Ariel Manzur.                  */
+/*                                                                        */
+/* Permission is hereby granted, free of charge, to any person obtaining  */
+/* a copy of this software and associated documentation files (the        */
+/* "Software"), to deal in the Software without restriction, including    */
+/* without limitation the rights to use, copy, modify, merge, publish,    */
+/* distribute, sublicense, and/or sell copies of the Software, and to     */
+/* permit persons to whom the Software is furnished to do so, subject to  */
+/* the following conditions:                                              */
+/*                                                                        */
+/* The above copyright notice and this permission notice shall be         */
+/* included in all copies or substantial portions of the Software.        */
+/*                                                                        */
+/* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,        */
+/* EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF     */
+/* MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. */
+/* IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY   */
+/* CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,   */
+/* TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE      */
+/* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                 */
+/**************************************************************************/
 
 #include "dir_access_unix.h"
 
-#if defined(UNIX_ENABLED) || defined(LIBC_FILEIO_ENABLED)
+#if defined(UNIX_ENABLED)
 
 #include "core/os/memory.h"
+#include "core/os/os.h"
 #include "core/string/print_string.h"
 #include "core/templates/list.h"
 
 #include <errno.h>
+#include <fcntl.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-
-#ifndef ANDROID_ENABLED
+#include <sys/ioctl.h>
 #include <sys/statvfs.h>
-#endif
 
 #ifdef HAVE_MNTENT
 #include <mntent.h>
 #endif
-
-DirAccess *DirAccessUnix::create_fs() {
-	return memnew(DirAccessUnix);
-}
 
 Error DirAccessUnix::list_dir_begin() {
 	list_dir_end(); //close any previous dir opening!
@@ -71,13 +67,13 @@ Error DirAccessUnix::list_dir_begin() {
 bool DirAccessUnix::file_exists(String p_file) {
 	GLOBAL_LOCK_FUNCTION
 
-	if (p_file.is_rel_path()) {
-		p_file = current_dir.plus_file(p_file);
+	if (p_file.is_relative_path()) {
+		p_file = current_dir.path_join(p_file);
 	}
 
 	p_file = fix_path(p_file);
 
-	struct stat flags;
+	struct stat flags = {};
 	bool success = (stat(p_file.utf8().get_data(), &flags) == 0);
 
 	if (success && S_ISDIR(flags.st_mode)) {
@@ -90,13 +86,13 @@ bool DirAccessUnix::file_exists(String p_file) {
 bool DirAccessUnix::dir_exists(String p_dir) {
 	GLOBAL_LOCK_FUNCTION
 
-	if (p_dir.is_rel_path()) {
-		p_dir = get_current_dir().plus_file(p_dir);
+	if (p_dir.is_relative_path()) {
+		p_dir = get_current_dir().path_join(p_dir);
 	}
 
 	p_dir = fix_path(p_dir);
 
-	struct stat flags;
+	struct stat flags = {};
 	bool success = (stat(p_dir.utf8().get_data(), &flags) == 0);
 
 	return (success && S_ISDIR(flags.st_mode));
@@ -105,8 +101,8 @@ bool DirAccessUnix::dir_exists(String p_dir) {
 bool DirAccessUnix::is_readable(String p_dir) {
 	GLOBAL_LOCK_FUNCTION
 
-	if (p_dir.is_rel_path()) {
-		p_dir = get_current_dir().plus_file(p_dir);
+	if (p_dir.is_relative_path()) {
+		p_dir = get_current_dir().path_join(p_dir);
 	}
 
 	p_dir = fix_path(p_dir);
@@ -116,8 +112,8 @@ bool DirAccessUnix::is_readable(String p_dir) {
 bool DirAccessUnix::is_writable(String p_dir) {
 	GLOBAL_LOCK_FUNCTION
 
-	if (p_dir.is_rel_path()) {
-		p_dir = get_current_dir().plus_file(p_dir);
+	if (p_dir.is_relative_path()) {
+		p_dir = get_current_dir().path_join(p_dir);
 	}
 
 	p_dir = fix_path(p_dir);
@@ -125,22 +121,22 @@ bool DirAccessUnix::is_writable(String p_dir) {
 }
 
 uint64_t DirAccessUnix::get_modified_time(String p_file) {
-	if (p_file.is_rel_path()) {
-		p_file = current_dir.plus_file(p_file);
+	if (p_file.is_relative_path()) {
+		p_file = current_dir.path_join(p_file);
 	}
 
 	p_file = fix_path(p_file);
 
-	struct stat flags;
+	struct stat flags = {};
 	bool success = (stat(p_file.utf8().get_data(), &flags) == 0);
 
 	if (success) {
 		return flags.st_mtime;
 	} else {
 		ERR_FAIL_V(0);
-	};
+	}
 	return 0;
-};
+}
 
 String DirAccessUnix::get_next() {
 	if (!dir_stream) {
@@ -162,9 +158,9 @@ String DirAccessUnix::get_next() {
 	// known if it points to a directory. stat() will resolve the link
 	// for us.
 	if (entry->d_type == DT_UNKNOWN || entry->d_type == DT_LNK) {
-		String f = current_dir.plus_file(fname);
+		String f = current_dir.path_join(fname);
 
-		struct stat flags;
+		struct stat flags = {};
 		if (stat(f.utf8().get_data(), &flags) == 0) {
 			_cisdir = S_ISDIR(flags.st_mode);
 		} else {
@@ -195,7 +191,7 @@ void DirAccessUnix::list_dir_end() {
 	_cisdir = false;
 }
 
-#if defined(HAVE_MNTENT) && defined(X11_ENABLED)
+#if defined(HAVE_MNTENT) && defined(LINUXBSD_ENABLED)
 static bool _filter_drive(struct mntent *mnt) {
 	// Ignore devices that don't point to /dev
 	if (strncmp(mnt->mnt_fsname, "/dev", 4) != 0) {
@@ -216,8 +212,11 @@ static bool _filter_drive(struct mntent *mnt) {
 #endif
 
 static void _get_drives(List<String> *list) {
-#if defined(HAVE_MNTENT) && defined(X11_ENABLED)
-	// Check /etc/mtab for the list of mounted partitions
+	// Add root.
+	list->push_back("/");
+
+#if defined(HAVE_MNTENT) && defined(LINUXBSD_ENABLED)
+	// Check /etc/mtab for the list of mounted partitions.
 	FILE *mtab = setmntent("/etc/mtab", "r");
 	if (mtab) {
 		struct mntent mnt;
@@ -237,7 +236,7 @@ static void _get_drives(List<String> *list) {
 	}
 #endif
 
-	// Add $HOME
+	// Add $HOME.
 	const char *home = getenv("HOME");
 	if (home) {
 		// Only add if it's not a duplicate
@@ -246,7 +245,8 @@ static void _get_drives(List<String> *list) {
 			list->push_back(home_name);
 		}
 
-		// Check $HOME/.config/gtk-3.0/bookmarks
+		// Check GTK+3 bookmarks for both XDG locations (Documents, Downloads, etc.)
+		// and potential user-defined bookmarks.
 		char path[1024];
 		snprintf(path, 1024, "%s/.config/gtk-3.0/bookmarks", home);
 		FILE *fd = fopen(path, "r");
@@ -255,7 +255,7 @@ static void _get_drives(List<String> *list) {
 			while (fgets(string, 1024, fd)) {
 				// Parse only file:// links
 				if (strncmp(string, "file://", 7) == 0) {
-					// Strip any unwanted edges on the strings and push_back if it's not a duplicate
+					// Strip any unwanted edges on the strings and push_back if it's not a duplicate.
 					String fpath = String::utf8(string + 7).strip_edges().split_spaces()[0].uri_decode();
 					if (!list->find(fpath)) {
 						list->push_back(fpath);
@@ -264,6 +264,12 @@ static void _get_drives(List<String> *list) {
 			}
 
 			fclose(fd);
+		}
+
+		// Add Desktop dir.
+		String dpath = OS::get_singleton()->get_system_dir(OS::SystemDir::SYSTEM_DIR_DESKTOP);
+		if (dpath.length() > 0 && !list->find(dpath)) {
+			list->push_back(dpath);
 		}
 	}
 
@@ -283,7 +289,21 @@ String DirAccessUnix::get_drive(int p_drive) {
 
 	ERR_FAIL_INDEX_V(p_drive, list.size(), "");
 
-	return list[p_drive];
+	return list.get(p_drive);
+}
+
+int DirAccessUnix::get_current_drive() {
+	int drive = 0;
+	int max_length = -1;
+	const String path = get_current_dir().to_lower();
+	for (int i = 0; i < get_drive_count(); i++) {
+		const String d = get_drive(i).to_lower();
+		if (max_length < d.length() && path.begins_with(d)) {
+			max_length = d.length();
+			drive = i;
+		}
+	}
+	return drive;
 }
 
 bool DirAccessUnix::drives_are_shortcuts() {
@@ -293,8 +313,8 @@ bool DirAccessUnix::drives_are_shortcuts() {
 Error DirAccessUnix::make_dir(String p_dir) {
 	GLOBAL_LOCK_FUNCTION
 
-	if (p_dir.is_rel_path()) {
-		p_dir = get_current_dir().plus_file(p_dir);
+	if (p_dir.is_relative_path()) {
+		p_dir = get_current_dir().path_join(p_dir);
 	}
 
 	p_dir = fix_path(p_dir);
@@ -304,11 +324,11 @@ Error DirAccessUnix::make_dir(String p_dir) {
 
 	if (success) {
 		return OK;
-	};
+	}
 
 	if (err == EEXIST) {
 		return ERR_ALREADY_EXISTS;
-	};
+	}
 
 	return ERR_CANT_CREATE;
 }
@@ -321,15 +341,15 @@ Error DirAccessUnix::change_dir(String p_dir) {
 	// prev_dir is the directory we are changing out of
 	String prev_dir;
 	char real_current_dir_name[2048];
-	ERR_FAIL_COND_V(getcwd(real_current_dir_name, 2048) == nullptr, ERR_BUG);
-	if (prev_dir.parse_utf8(real_current_dir_name)) {
+	ERR_FAIL_NULL_V(getcwd(real_current_dir_name, 2048), ERR_BUG);
+	if (prev_dir.parse_utf8(real_current_dir_name) != OK) {
 		prev_dir = real_current_dir_name; //no utf8, maybe latin?
 	}
 
 	// try_dir is the directory we are trying to change into
 	String try_dir = "";
-	if (p_dir.is_rel_path()) {
-		String next_dir = current_dir.plus_file(p_dir);
+	if (p_dir.is_relative_path()) {
+		String next_dir = current_dir.path_join(p_dir);
 		next_dir = next_dir.simplify_path();
 		try_dir = next_dir;
 	} else {
@@ -342,8 +362,8 @@ Error DirAccessUnix::change_dir(String p_dir) {
 	}
 
 	String base = _get_root_path();
-	if (base != String() && !try_dir.begins_with(base)) {
-		ERR_FAIL_COND_V(getcwd(real_current_dir_name, 2048) == nullptr, ERR_BUG);
+	if (!base.is_empty() && !try_dir.begins_with(base)) {
+		ERR_FAIL_NULL_V(getcwd(real_current_dir_name, 2048), ERR_BUG);
 		String new_dir;
 		new_dir.parse_utf8(real_current_dir_name);
 
@@ -358,9 +378,9 @@ Error DirAccessUnix::change_dir(String p_dir) {
 	return OK;
 }
 
-String DirAccessUnix::get_current_dir(bool p_include_drive) {
+String DirAccessUnix::get_current_dir(bool p_include_drive) const {
 	String base = _get_root_path();
-	if (base != "") {
+	if (!base.is_empty()) {
 		String bd = current_dir.replace_first(base, "");
 		if (bd.begins_with("/")) {
 			return _get_root_string() + bd.substr(1, bd.length());
@@ -372,14 +392,14 @@ String DirAccessUnix::get_current_dir(bool p_include_drive) {
 }
 
 Error DirAccessUnix::rename(String p_path, String p_new_path) {
-	if (p_path.is_rel_path()) {
-		p_path = get_current_dir().plus_file(p_path);
+	if (p_path.is_relative_path()) {
+		p_path = get_current_dir().path_join(p_path);
 	}
 
 	p_path = fix_path(p_path);
 
-	if (p_new_path.is_rel_path()) {
-		p_new_path = get_current_dir().plus_file(p_new_path);
+	if (p_new_path.is_relative_path()) {
+		p_new_path = get_current_dir().path_join(p_new_path);
 	}
 
 	p_new_path = fix_path(p_new_path);
@@ -388,18 +408,18 @@ Error DirAccessUnix::rename(String p_path, String p_new_path) {
 }
 
 Error DirAccessUnix::remove(String p_path) {
-	if (p_path.is_rel_path()) {
-		p_path = get_current_dir().plus_file(p_path);
+	if (p_path.is_relative_path()) {
+		p_path = get_current_dir().path_join(p_path);
 	}
 
 	p_path = fix_path(p_path);
 
-	struct stat flags;
+	struct stat flags = {};
 	if ((stat(p_path.utf8().get_data(), &flags) != 0)) {
 		return FAILED;
 	}
 
-	if (S_ISDIR(flags.st_mode)) {
+	if (S_ISDIR(flags.st_mode) && !is_link(p_path)) {
 		return ::rmdir(p_path.utf8().get_data()) == 0 ? OK : FAILED;
 	} else {
 		return ::unlink(p_path.utf8().get_data()) == 0 ? OK : FAILED;
@@ -407,23 +427,23 @@ Error DirAccessUnix::remove(String p_path) {
 }
 
 bool DirAccessUnix::is_link(String p_file) {
-	if (p_file.is_rel_path()) {
-		p_file = get_current_dir().plus_file(p_file);
+	if (p_file.is_relative_path()) {
+		p_file = get_current_dir().path_join(p_file);
 	}
 
 	p_file = fix_path(p_file);
 
-	struct stat flags;
+	struct stat flags = {};
 	if ((lstat(p_file.utf8().get_data(), &flags) != 0)) {
-		return FAILED;
+		return false;
 	}
 
 	return S_ISLNK(flags.st_mode);
 }
 
 String DirAccessUnix::read_link(String p_file) {
-	if (p_file.is_rel_path()) {
-		p_file = get_current_dir().plus_file(p_file);
+	if (p_file.is_relative_path()) {
+		p_file = get_current_dir().path_join(p_file);
 	}
 
 	p_file = fix_path(p_file);
@@ -439,8 +459,8 @@ String DirAccessUnix::read_link(String p_file) {
 }
 
 Error DirAccessUnix::create_link(String p_source, String p_target) {
-	if (p_target.is_rel_path()) {
-		p_target = get_current_dir().plus_file(p_target);
+	if (p_target.is_relative_path()) {
+		p_target = get_current_dir().path_join(p_target);
 	}
 
 	p_source = fix_path(p_source);
@@ -454,18 +474,13 @@ Error DirAccessUnix::create_link(String p_source, String p_target) {
 }
 
 uint64_t DirAccessUnix::get_space_left() {
-#ifndef NO_STATVFS
 	struct statvfs vfs;
 	if (statvfs(current_dir.utf8().get_data(), &vfs) != 0) {
 		return 0;
-	};
+	}
 
 	return (uint64_t)vfs.f_bavail * (uint64_t)vfs.f_frsize;
-#else
-	// FIXME: Implement this.
-	return 0;
-#endif
-};
+}
 
 String DirAccessUnix::get_filesystem_type() const {
 	return ""; //TODO this should be implemented
@@ -473,6 +488,27 @@ String DirAccessUnix::get_filesystem_type() const {
 
 bool DirAccessUnix::is_hidden(const String &p_name) {
 	return p_name != "." && p_name != ".." && p_name.begins_with(".");
+}
+
+bool DirAccessUnix::is_case_sensitive(const String &p_path) const {
+#if defined(LINUXBSD_ENABLED)
+	String f = p_path;
+	if (!f.is_absolute_path()) {
+		f = get_current_dir().path_join(f);
+	}
+	f = fix_path(f);
+
+	int fd = ::open(f.utf8().get_data(), O_RDONLY | O_NONBLOCK);
+	if (fd) {
+		long flags = 0;
+		if (ioctl(fd, _IOR('f', 1, long), &flags) >= 0) {
+			::close(fd);
+			return !(flags & 0x40000000 /* FS_CASEFOLD_FL */);
+		}
+		::close(fd);
+	}
+#endif
+	return true;
 }
 
 DirAccessUnix::DirAccessUnix() {
@@ -483,8 +519,8 @@ DirAccessUnix::DirAccessUnix() {
 
 	// set current directory to an absolute path of the current directory
 	char real_current_dir_name[2048];
-	ERR_FAIL_COND(getcwd(real_current_dir_name, 2048) == nullptr);
-	if (current_dir.parse_utf8(real_current_dir_name)) {
+	ERR_FAIL_NULL(getcwd(real_current_dir_name, 2048));
+	if (current_dir.parse_utf8(real_current_dir_name) != OK) {
 		current_dir = real_current_dir_name;
 	}
 
@@ -495,4 +531,4 @@ DirAccessUnix::~DirAccessUnix() {
 	list_dir_end();
 }
 
-#endif //posix_enabled
+#endif // UNIX_ENABLED
